requirements.txt
paramiko==2.0.0 # LGPLv2.1+
pytest>=3.0.4  # MIT
python-cinderclient>=1.6.0,!=1.7.0,!=1.7.1  # Apache-2.0
python-glanceclient>=2.5.0  # Apache-2.0
python-keystoneclient>=3.8.0  # Apache-2.0
python-neutronclient>=5.1.0  # Apache-2.0
python-novaclient==7.1.0
PyYAML>=3.12  # MIT
requests>=2.10.0,!=2.12.2 # Apache-2.0
texttable
